# Proyecto/E3 (Final)/BACKEND/swagger.yaml
openapi: 3.0.0
info:
  title: Dawdle API
  description: |
    API para gestión de eventos y usuarios
    
    ## Autenticación
    La API utiliza un sistema híbrido de autenticación:
    - **JWT (JSON Web Token)**: Se devuelve en el cuerpo de la respuesta con expiración de 1 hora
    - **Cookies httpOnly**: Se establece automáticamente en las rutas `/auth/register` y `/auth/login` con las siguientes características:
      - `httpOnly: true` - No accesible desde JavaScript (protección contra XSS)
      - `secure: true` - Solo en HTTPS (producción)
      - `sameSite: 'lax'` - Protección contra CSRF
      - `maxAge: 3600000` - 1 hora (3600 segundos)
      - `path: '/'` - Disponible en toda la aplicación
    
    ## Seguridad de contraseñas
    Las contraseñas se almacenan utilizando bcrypt con **12 rounds de salt** para máxima seguridad.
    
    ## CORS
    El servidor acepta peticiones solo desde el origen configurado en `FRONT_ORIGIN` (variable de entorno).
  version: 1.0.0
  contact:
    name: Soporte Dawdle
    email: soporte@dawdle.com
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://dawdle-backend.onrender.com
    description: Servidor de producción
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Obtener todos los usuarios
      description: Retorna lista de usuarios (requiere ser admin)
      responses:
        '200':
          description: Lista de usuarios obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: No tiene permisos de administrador
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Se requieren permisos de administrador."
  /users/{id}:
    put:
      tags:
        - Users
      summary: Actualizar usuario
      description: Actualizar información de usuario (solo dueño o admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario actualizado exitosamente"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "La edad debe ser un número válido"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: Sin permisos para editar este usuario o username reservado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "El username 'admin' está reservado y no puede ser usado"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Usuario no encontrado"
        '409':
          description: Username o email ya en uso
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "El username ya está en uso"
    delete:
      tags:
        - Users
      summary: Eliminar usuario
      description: Eliminar usuario (solo admin o dueño)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Usuario eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario eliminado exitosamente"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: Sin permisos para eliminar este usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para eliminar este usuario"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Usuario no encontrado"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "nuevo_usuario"
        email:
          type: string
          format: email
          example: "usuario@example.com"
        image:
          type: string
          format: uri
          example: "https://ui-avatars.com/api/?name=nuevo_usuario&background=random"
        age:
          type: integer
          example: 25
        birthday:
          type: string
          format: date-time
          example: "1999-05-20T00:00:00.000Z"
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          example: "nuevo_usuario_cambiado"
        email:
          type: string
          format: email
          example: "usuario@example.com"
        age:
          type: integer
          example: 30
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []